fix(sandbox): bake project source synchronizer
